% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitbit_activity_data.R
\name{extract_fitbit_activity_data}
\alias{extract_fitbit_activity_data}
\title{Extract activity data}
\usage{
extract_fitbit_activity_data(csvFileName)
}
\arguments{
\item{csvFileName}{The path to the .csv file with activity data}
}
\value{
a tibble
}
\description{
Extract the activity data from a very messy "csv" file exported
from the FitBit web site
}
\examples{

}
