% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fitbit_sleep_data.R
\name{extract_fitbit_sleep_data}
\alias{extract_fitbit_sleep_data}
\title{Extract Fitbit sleep data}
\usage{
extract_fitbit_sleep_data(csvFileName, bVerbose = FALSE)
}
\arguments{
\item{csvFileName}{The path to the .csv file with activity data}

\item{bVerbose}{Print warnings if TRUE. Default: False}
}
\value{
a tibble
}
\description{
Extract the sleep data from a very messy "csv" file exported
from the FitBit web site
}
\examples{

}
